Fixed 64-bit related issues

+++ demos/new/dgl/asset/dgl2.d
@@ -204,7 +204,7 @@ void loadDGL2(InputStream istrm, Scene scene)
         else if (chunk.type == ChunkType.TRIMESH)
         {
             assert(!(chunk.data.length % DGLTriangle.sizeof)); // Check data integrity
-            uint numTris = chunk.data.length / DGLTriangle.sizeof;
+            size_t numTris = chunk.data.length / DGLTriangle.sizeof;
             Triangle[] tris = New!(Triangle[])(numTris);
             auto mtris = cast(DGLTriangle[])chunk.data;
             uint offset = 0;
